requirements.txt
requests>=2.31.0
cloudscraper>=1.2.71
httpx[http2]>=0.27.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
ebooklib>=0.18
python-docx>=1.0.0
reportlab>=4.0.0
Pillow>=10.0.0
